requirements.txt
numpy>=1.16.2
torch==1.5
torchvision==0.6.0
torchgeometry>=0.1.2
PyYAML==5.1.1
configargparse
smplx==0.1.15
pillow
opencv-python
tqdm
yacs
ipython
ipdb
scipy
trimesh
pyrender==0.1.36
kornia==0.2.0
tensorboardX
matplotlib
easydict
joblib
